requirements.txt
PyYAML
pymysql
allure-pytest== 2.8.16
allure-python-commons==2.8.16
chardet== 3.0.4
colorama== 0.4.3
colorlog== 4.1.0
filetype== 1.0.7
flower== 0.9.2
idna== 2.9
lxml== 4.5.1
mysqlclient== 1.4.6
paramiko== 2.4.1
pyasn1== 0.4.8
pyparsing==2.4.7
pytest== 5.1.3
pytest-html== 2.1.1
pytest-metadata== 1.9.0
requests== 2.23.0
requests-toolbelt==	0.9.1
scikit-learn==0.23.1
scipy==1.5.0
setuptools== 41.2.0
sqlparse== 0.3.1
synonyms== 3.10.2
tornado==6.0.4
urllib3==1.25.9